[RENAME] 변경된 경로명 적용하기 #NADA-52
