+++ src/components/common/filter/Dropdown.js
@@ -2,10 +2,10 @@
 import PropTypes from 'prop-types';
 import { useState, useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
-import { filterSelector, setField } from '../../modules/filter';
-import { Dropdown, TextWarpper } from '../../styles/community/DropdownStyle';
-import { selectAllConfig } from '../common/AttributeData';
-import { FilterHandler } from '../common/icon/FilterHandler';
+import { filterSelector, setField } from '../../../modules/filter';
+import { Dropdown, TextWarpper } from '../../../styles/community/DropdownStyle';
+import { selectAllConfig } from '../../common/AttributeData';
+import { FilterHandler } from '../../common/icon/FilterHandler';
 
 export const DropDown = ({ content, text }) => {
   const dispatch = useDispatch();
