+++ src/components/common/modal/ShareModal.js
@@ -1,6 +1,6 @@
 /** ShareModal 공유 방법을 선택하는 모달 */
 import { Layout } from '../../../styles/community/NoticeModalStyle';
-import ModalButtonDiv from './ModalButtonDiv';
+import ModalButtonDiv from '../usedInModal/ModalButtonDiv';
 
 const ShareModal = () => {
   return (
